refactor front end book and ownership creation
