+++ app/views/books/new.html.erb
@@ -25,25 +25,25 @@
       </div>
       <div class="modal-body">
         <%= form_for [@book], multipart: true do |f| %>
-        <div class="form-group">
-          <%= f.label :title%>
-          <%= f.text_field :title, class: "form-control text-center" %>
-        </div>
+          <div class="form-group">
+            <%= f.label :title%>
+            <%= f.text_field :title, class: "form-control text-center" %>
+          </div>
 
-        <div class="form-group">
-          <%= f.label :author%>
-          <%= f.text_field :author, class: "form-control text-center" %>
-        </div>
+          <div class="form-group">
+            <%= f.label :author%>
+            <%= f.text_field :author, class: "form-control text-center" %>
+          </div>
 
-        <div class="form-group">
-          <%= f.label :cover %><br/>
-          <%= f.file_field :cover %>
-          <p class="help-block">Upload a cover.</p>
-        </div>
+          <div class="form-group">
+            <%= f.label :cover %><br/>
+            <%= f.file_field :cover %>
+            <p class="help-block">Upload a cover.</p>
+          </div>
 
-        <div class="form-group">
-          <%= f.submit "Submit", class: "btn btn-primary btn-block" %>
-        </div>
+          <div class="form-group">
+            <%= f.submit "Submit", class: "btn btn-primary btn-block" %>
+          </div>
         <% end %>
       </div>
     </div>
@@ -84,5 +84,8 @@
 <!-- /.modal -->
 
 <script type="text/javascript">
-window.bookPredictor.init('#title');
+  var bookId = -1;
+  window.bookPredictor.init('#title');
+  window.bookNewManager.init();
+  window.bookOwnershipManager.init();
 </script>
