+++ app/views/ownerships/_ownership_list.html.erb
@@ -28,6 +28,6 @@
       <% end %>
     <% end %>
   <% else %>
-  <p class="lead text-center">No books offered yet!</p>
+    <p class="lead text-center">No books offered yet!</p>
   <% end %>
 </div>
